Coerce lists and character vectors to OAuth2 implicit flows

as_oauth2_implicit_flow() turns an existing object into an oauth2_implicit_flow. This is in contrast with class_oauth2_implicit_flow(), which builds an oauth2_implicit_flow from individual properties.

Usage

as_oauth2_implicit_flow(x, ..., arg = caller_arg(x), call = caller_env())

Arguments

x

The object to coerce. Must be empty or be a list of named lists, each with names "refresh_url", "scopes", and/or "authorization_url", or names that can be coerced to those names via snakecase::to_snake_case(). Additional names are ignored.

...

These dots are for future extensions and must be empty.

arg

An argument name as a string. This argument will be mentioned in error messages as the input that is at the origin of a problem.

call

The execution environment of a currently running function, e.g. caller_env(). The function will be mentioned in error messages as the source of the error. See the call argument of abort() for more information.

Value

An oauth2_implicit_flow as returned by class_oauth2_implicit_flow().

See also