API key security schemes

Defines an API key security scheme that can be used by the operations.

Usage

class_api_key_security_scheme(
  parameter_name = character(),
  location = c("query", "header", "cookie")
)

Arguments

parameter_name

Character vector (required). The names of the header, query or cookie parameters to be used.

location

Character vector (required). The location of the API key. Valid values are "query", "header" or "cookie".

Value

An api_key_security_scheme S7 object, with fields parameter_name and location.

See also

Examples

class_api_key_security_scheme(
  parameter_name = "Authorization",
  location = "header"
)
#> <rapid::api_key_security_scheme>
#>  @ parameter_name: chr "Authorization"
#>  @ location      : chr "header"