OAuth2 security schemes

Defines an OAuth2 security scheme that can be used by the operations.

Usage

class_oauth2_security_scheme(
  ...,
  implicit_flow = class_oauth2_implicit_flow(),
  password_flow = class_oauth2_token_flow(),
  client_credentials_flow = class_oauth2_token_flow(),
  authorization_code_flow = class_oauth2_authorization_code_flow()
)

Arguments

...

These dots are for future extensions and must be empty.

implicit_flow

An oauth2_implicit_flow object created with class_oauth2_implicit_flow().

password_flow, client_credentials_flow

An oauth2_token_flow object created with class_oauth2_token_flow().

authorization_code_flow

An oauth2_authorization_code_flow object created with class_oauth2_authorization_code_flow().

Value

An oauth2_security_scheme S7 object, with fields implicit_flow, password_flow, client_credentials_flow, and authorization_code_flow.

Examples

class_oauth2_security_scheme()
#> <rapid::oauth2_security_scheme>
#>  @ implicit_flow          : <rapid::oauth2_implicit_flow>
#>  .. @ refresh_url      : chr(0) 
#>  .. @ scopes           : <rapid::scopes>
#>  .. .. @ name       : chr(0) 
#>  .. .. @ description: chr(0) 
#>  .. @ authorization_url: chr(0) 
#>  @ password_flow          : <rapid::oauth2_token_flow>
#>  .. @ refresh_url: chr(0) 
#>  .. @ scopes     : <rapid::scopes>
#>  .. .. @ name       : chr(0) 
#>  .. .. @ description: chr(0) 
#>  .. @ token_url  : chr(0) 
#>  @ client_credentials_flow: <rapid::oauth2_token_flow>
#>  .. @ refresh_url: chr(0) 
#>  .. @ scopes     : <rapid::scopes>
#>  .. .. @ name       : chr(0) 
#>  .. .. @ description: chr(0) 
#>  .. @ token_url  : chr(0) 
#>  @ authorization_code_flow: <rapid::oauth2_authorization_code_flow>
#>  .. @ refresh_url      : chr(0) 
#>  .. @ scopes           : <rapid::scopes>
#>  .. .. @ name       : chr(0) 
#>  .. .. @ description: chr(0) 
#>  .. @ authorization_url: chr(0) 
#>  .. @ token_url        : chr(0) 
class_oauth2_security_scheme(
  password_flow = class_oauth2_token_flow(token_url = "/tokens/passwords")
)
#> <rapid::oauth2_security_scheme>
#>  @ implicit_flow          : <rapid::oauth2_implicit_flow>
#>  .. @ refresh_url      : chr(0) 
#>  .. @ scopes           : <rapid::scopes>
#>  .. .. @ name       : chr(0) 
#>  .. .. @ description: chr(0) 
#>  .. @ authorization_url: chr(0) 
#>  @ password_flow          : <rapid::oauth2_token_flow>
#>  .. @ refresh_url: chr(0) 
#>  .. @ scopes     : <rapid::scopes>
#>  .. .. @ name       : chr(0) 
#>  .. .. @ description: chr(0) 
#>  .. @ token_url  : chr "/tokens/passwords"
#>  @ client_credentials_flow: <rapid::oauth2_token_flow>
#>  .. @ refresh_url: chr(0) 
#>  .. @ scopes     : <rapid::scopes>
#>  .. .. @ name       : chr(0) 
#>  .. .. @ description: chr(0) 
#>  .. @ token_url  : chr(0) 
#>  @ authorization_code_flow: <rapid::oauth2_authorization_code_flow>
#>  .. @ refresh_url      : chr(0) 
#>  .. @ scopes           : <rapid::scopes>
#>  .. .. @ name       : chr(0) 
#>  .. .. @ description: chr(0) 
#>  .. @ authorization_url: chr(0) 
#>  .. @ token_url        : chr(0)