OAuth2 flow scopes objects

The available scopes for an OAuth2 flow.

Usage

class_scopes(name = character(), description = character())

Arguments

name

Character vector (required). The name of the scope.

description

Character vector (required). A short description of the scope.

Value

A scopes S7 object, with fields name and description.

See also

Examples

class_scopes(
  name = c(
    "https://www.googleapis.com/auth/youtube",
    "https://www.googleapis.com/auth/youtube.channel-memberships.creator",
    "https://www.googleapis.com/auth/youtube.force-ssl",
    "https://www.googleapis.com/auth/youtube.readonly",
    "https://www.googleapis.com/auth/youtube.upload",
    "https://www.googleapis.com/auth/youtubepartner",
    "https://www.googleapis.com/auth/youtubepartner-channel-audit"
  ),
  description = c(
    "Manage your YouTube account",
    "See a list of your current active channel members",
    "See, edit, and permanently delete your YouTube videos",
    "View your YouTube account",
    "Manage your YouTube videos",
    "View and manage your assets and associated content on YouTube",
    "View private information of your YouTube channel"
  )
)
#> <rapid::scopes>
#>  @ name       : chr [1:7] "https://www.googleapis.com/auth/youtube" ...
#>  @ description: chr [1:7] "Manage your YouTube account" ...